Update Query database endpoint to return both Page and Database
