+++ Src/Notion.Client/Api/Databases/Query/Response/DatabaseQueryResponse.cs
@@ -1,0 +1,7 @@
+﻿namespace Notion.Client
+{
+    // ReSharper disable once ClassNeverInstantiated.Global
+    public class DatabaseQueryResponse : PaginatedList<IWikiDatabase>
+    {
+    }
+}
